requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-groq==0.3.8

# Embeddings & Vector Store
sentence-transformers==2.2.2
faiss-cpu==1.13.2

# Frontend
streamlit==1.50.0

# Optional
numpy>=1.24.0
